+++ nextgisweb/resource/nodepkg/description-editor/index.js
@@ -1,0 +1,4 @@
+/** @entrypoint */
+import { DescriptionEditorWidget } from "./DescriptionEditorWidget";
+import { DescriptionEditorStore } from "./DescriptionEditorStore";
+export { DescriptionEditorStore as store, DescriptionEditorWidget as widget };